refactor: Fix generics and Bootstrapping
Fixed use of generics in DestinationService and SourceService. Compiler can't guarantee the explicit extension of Destination (or Source), so instead just use ? and the interface as return types. The code to grab the right repository still needs an explicit generic, but otherwise use this pattern. This allows the bootstrapping code to do away with explicitly casting to each potential type in turn.