feat: FOLIO Authentication
Initial attempt at dealing with FOLIO's authentication. Creates a CookieToken, held in the FOLIODestinationApiClient, which gives us neat methods and allows us to send and recieve auth cookies to talk to FOLIO.

Username/Password now stored in db (erk) in FolioDestination, Bootstrapped in (Dummy data so as not to store credentials in gitlab)

Working with the option to fetch all agreements from "/erm/sas" -- this will be removed eventually.