feat: Gokb
Added Gokb object which holds the baseUrl. This can then be SHARED across Gokb sources, ie TIPP vs Package.

Not quite as useful here, but could be _very_useful as a pattern to allow us to hold FOLIO tenant level information for both source/destination in a single place, but leave the destinations as separate (PCI in vs PKG in, say)